+++ changelog.md
@@ -1,4 +1,34 @@
 ## 更新日志 Changelog
+- V5.5.10 2024-12-01
+  - CMD/Ctrl+A 增强功能支持quote/callout，第一次选中当前引用行内容，第二次选中当前的整个引用 #260
+  - 切换代码块注释命令
+    - 修复对 CSS 代码块注释错误的问题 #259
+    - 增加了对 HTML、Markdown、Matlab 代码块注释的支持
+    - 优化该命令在空的代码行的表现
+  - 修复 MD 链接内部有小括号时会被错误格式化的问题 #178
+  - Enhanced functionality for CMD/Ctrl+A to support quote/callout, the first press selects the current quote line content, the second press selects the entire current quote #260
+  - Toggle code block comment command
+    - Fixed the issue of incorrect comments in CSS code blocks #259
+    - Added support for comments in HTML, Markdown, and Matlab code blocks
+    - Optimized the performance of this command on empty code lines
+  - Fixed the issue where MD links with parentheses inside would be incorrectly formatted #178
+- V5.5.9 2024-11-23
+  - 严格换行模式支持三种模式（两次换行，两次空格+换行，混合模式）#193
+    - 双空格模式：回车会变成两次空格+换行
+    - 混合模式是在引用块中使用两次空格+换行，在其他地方使用两次换行
+  - 修复了列表中代码块结尾回车时，触发两次换行的问题
+  - 增加了选中当前文本块的快捷键命令 #256
+  - 修改了选中文本块的逻辑，目前不需要严格换行模式也把相邻的文本行作为同一文本块的内容 #255。
+- V5.5.8 2024-11-19
+  - 紧急修复上次更新导致的一个输入问题
+  - Urgently fixed an input issue caused by the last update
+- V5.5.7 2024-11-19
+  - 严格换行两次支持 quote 块 #254
+  - 增加了引用符号 > 与文本之间自动空格的功能，默认开启，可以在设置中关闭
+  - 优化了在基础输入增强中，在句首输入 》或者 > 的行为
+  - Added support for strict line breaks twice in quote blocks #254
+  - Added the function of automatically inserting a space between the quote symbol > and the text, enabled by default, can be turned off in the settings
+  - Optimized the behavior of inputting 》or > at the beginning of a sentence in basic input enhancement
 - V5.5.6 2024-11-14
   - 增加了 Ctrl/Cmd+A 增强功能，第一次选中当前行，第二次选中当前文本块，第三次选中全文 #255
     - (目前仅在纯文本块中生效，不包括引用和列表)